edit monitoring tabpane
